requirements.txt
# Financial Risk Assessment System Dependencies
# Core GUI Framework
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# Database Support
psycopg2-binary==2.9.9
sqlite3

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Security
bcrypt==4.1.2
cryptography==41.0.8

# Logging and Monitoring
loguru==0.7.2

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-qt==4.2.0
pytest-cov==4.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development
pre-commit==3.6.0 